API/requirements.txt
streamlit
requests
scikit-learn
flask
pillow
tensorflow
numpy
